+++ experimental/CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to experimental packages in this project will be documented 
 
 ### :bug: (Bug Fix)
 
+* fix(instrumentation-http): Ensure instrumentation of `http.get` and `https.get` work when used in ESM code [#4857](https://github.com/open-telemetry/opentelemetry-js/issues/4857) @trentm
+
 ### :books: (Refine Doc)
 
 ### :house: (Internal)